Improve function

+++ cpp_text_shuffle/function.cpp
@@ -6,8 +6,8 @@ using std::cout;
 
 string text_shuffle(string text){
   srand(time(NULL));
-  for (int i=0; i<(text.size())/2; i++){
-    unsigned int random_index = rand()%text.size();
+  for (uintmax_t i=0; i<(text.size())/2; i++){
+    uintmax_t random_index = rand()%text.size();
     text.push_back(text[random_index]);
     text.erase(random_index, 1);
   }
